requirements.txt
tk
face_recognition
mysql.connector
Pillow
opencv_python
numpy 